requirements.txt
# 核心依赖
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pyyaml>=6.0

# 通义千问SDK
dashscope>=1.14.0

# 数据处理
python-dateutil>=2.8.0

# 可选：加速解析
html5lib>=1.1
